nlp-service/requirements.txt
kafka-python
spacy
sentence-transformers
torch
transformers
numpy
pandas
minio
pytesseract
Pillow
requests
pymupdf
